sdk/requirements.txt
azure-ai-ml
python-dotenv